+++ CHANGELOG.md
@@ -30,5 +30,6 @@
 ## `1.0.0-2`
 
 - Support markdown descriptions for collectibles and bios
+- Fixup ID setup not updating addresses
 - Fix 404 errors incorrectly displayed on Orbit
 - Custom 500 error page